mcp-nano-banana/pyproject.toml
[project]
name = "mcp-nano-banana"
version = "1.0.0"
description = "Diagram generation MCP server — best-in-class HTML/SVG diagrams for presentations"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "MIT"}
authors = [{name = "cooneycw"}]

dependencies = [
    "mcp[cli]>=1.2.0",
    "fastmcp>=1.0",
    "pydantic>=2.6.0",
    "python-pptx>=1.0.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]
